Initial contents of main window
The main window now has catalog and account tabs, a logout button and
user label.
The tabs dont do anything yet but the account one will only show up if
the current account is a staff account.
Logout button brings the login dialog back up which is then able to
restructure the main window accordingly upon re-login.

The panels for the tabs in the main window need to be custom panels so
that the MainFrame.java doesnt become a massive mess of GUI bullshit