Account functionality added
Cleaned up main window.
Catalog and accounts are no longer in tabs because it was screwing with
the table updating. Catalog is now the main frame, and the accounts will
open in its own smaller frame.

Main frame was visually updated to be a little more user friendly (i
think)