Checkout added but not finished
created the checkout frame, which contains two tables and some buttons.

The frame is connected to a cart class, which holds a list of the books
being processed, along with a collection of observers that will address
the database changes upon checkout.

checking out seems to be no problem, but it seems that returning causing
some explosions once in a while, need to investigate.

also need to make sure that books are not allowed to be deleted if they
are checked out